moved image interaction functions to a .js file
